Implement dealing with stopwords in queries and an important fix to a bug which could cause results to be deleted before they had finished being used
A union query over any stop words is treated as a stop query. An intersection query with only one non-stopword is treated as a single word query on that word. An exclusion query with any stopwords is a stop query. A phrase query with more than one non-stopword and stopwords inbetween treats the stopwords as blanks when searching.