Avoid the use of the "freenet:"-prefix.
In some cases, when the FreenetArchiver works from an URI instead of from
an encoded number, files are created with the URI name. This does not work
on windows because colon ":" is not allowed by the file system.

This fix creates and uses the filenames without the freenet:-prefix in
those cases, not changing the more common case.

This is an attempt to fix https://bugs.freenetproject.org/view.php?id=5986.